requirements.txt
streamlit==1.28.0
tensorflow==2.14.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.0.1
pandas==2.1.1
plotly==5.17.0
